gitserver: Remove duplicative test
This test is now covered by gitcli and grpc server tests, so we no longer need this test here.

Test plan:

CI, code review.
